Split the Control component from the overall screen
For readability purposes.
