+++ app/src/main/kotlin/dev/shadoe/delta/control/ControlScreen.kt
@@ -1,6 +1,5 @@
 package dev.shadoe.delta.control
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -11,26 +10,20 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.ContentCopy
 import androidx.compose.material.icons.rounded.KeyboardArrowDown
 import androidx.compose.material.icons.rounded.KeyboardArrowUp
-import androidx.compose.material.icons.rounded.Password
-import androidx.compose.material.icons.rounded.QrCode2
-import androidx.compose.material.icons.rounded.Wifi
 import androidx.compose.material3.Button
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -40,21 +33,19 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalClipboardManager
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.window.core.layout.WindowWidthSizeClass
 import dev.shadoe.delta.R
 import dev.shadoe.delta.api.ACLDevice
 import dev.shadoe.delta.api.SoftApEnabledState
 import dev.shadoe.delta.control.components.AppBarWithDebugAction
-import dev.shadoe.delta.control.components.PassphraseDisplay
-import dev.shadoe.delta.control.components.SoftApControlButton
+import dev.shadoe.delta.control.components.SoftApControl
+import dev.shadoe.delta.control.components.SoftApControlViewModel
 import kotlinx.coroutines.launch
 
 @Composable
@@ -63,7 +54,6 @@ fun ControlScreen(
   modifier: Modifier = Modifier,
   vm: ControlViewModel = viewModel(),
 ) {
-  val context = LocalContext.current
   val clipboardManager = LocalClipboardManager.current
   val density = LocalDensity.current
 
@@ -71,13 +61,6 @@ fun ControlScreen(
   val snackbarHostState = remember { SnackbarHostState() }
   var isBlocklistShown by remember { mutableStateOf(false) }
 
-  val isBigScreen =
-    currentWindowAdaptiveInfo().windowSizeClass.windowWidthSizeClass ==
-      WindowWidthSizeClass.EXPANDED
-  val ssid by vm.ssid.collectAsState("")
-  val passphrase by vm.passphrase.collectAsState("")
-  val shouldShowPassphrase by vm.shouldShowPassphrase.collectAsState(true)
-  val shouldShowQrButton by vm.shouldShowQrButton.collectAsState(false)
   val enabledState by
     vm.enabledState.collectAsState(SoftApEnabledState.WIFI_AP_STATE_DISABLED)
   val tetheredClients by vm.connectedClients.collectAsState(listOf())
@@ -99,72 +82,19 @@ fun ControlScreen(
   ) { scaffoldPadding ->
     LazyColumn(modifier = Modifier.padding(scaffoldPadding)) {
       item {
-        Row(
-          modifier =
-            Modifier.fillMaxWidth()
-              .padding(horizontal = 16.dp)
-              .clip(RoundedCornerShape(8.dp))
-              .background(MaterialTheme.colorScheme.primaryContainer),
-          verticalAlignment = Alignment.CenterVertically,
-          horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-          Column(
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
-            verticalArrangement = Arrangement.Center,
-          ) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-              Icon(
-                imageVector = Icons.Rounded.Wifi,
-                contentDescription = stringResource(R.string.ssid_field_icon),
-                modifier = Modifier.padding(end = 8.dp),
+        val compVm = hiltViewModel<SoftApControlViewModel>()
+        SoftApControl(
+          onFailedToShowQr = {
+            scope.launch {
+              snackbarHostState.showSnackbar(
+                message = stringFeatureNotSupported,
+                withDismissAction = true,
+                duration = SnackbarDuration.Short,
               )
-              Text(text = ssid ?: stringResource(id = R.string.no_ssid))
             }
-            if (shouldShowPassphrase) {
-              Row(
-                modifier = Modifier.padding(vertical = 8.dp),
-                verticalAlignment = Alignment.CenterVertically,
-              ) {
-                Icon(
-                  imageVector = Icons.Rounded.Password,
-                  contentDescription =
-                    stringResource(R.string.passphrase_field_icon),
-                  modifier = Modifier.padding(end = 8.dp),
-                )
-                PassphraseDisplay(passphrase = passphrase)
-              }
-            }
-          }
-          Row(verticalAlignment = Alignment.CenterVertically) {
-            if (shouldShowQrButton) {
-              IconButton(
-                onClick = {
-                  if (!vm.openQrCodeScreen(context, isBigScreen)) {
-                    scope.launch {
-                      snackbarHostState.showSnackbar(
-                        message = stringFeatureNotSupported,
-                        withDismissAction = true,
-                        duration = SnackbarDuration.Short,
-                      )
-                    }
-                  }
-                }
-              ) {
-                Icon(
-                  imageVector = Icons.Rounded.QrCode2,
-                  contentDescription =
-                    stringResource(id = R.string.qr_code_button),
-                  modifier = Modifier.size(48.dp),
-                )
-              }
-            }
-            SoftApControlButton(
-              enabledState,
-              startHotspot = { vm.startHotspot() },
-              stopHotspot = { vm.stopHotspot() },
-            )
-          }
-        }
+          },
+          vm = compVm,
+        )
       }
       item {
         Text(
